Изменил логику загрузки предложения, подключил предложения неподалеку, порефакторил код
